requirements.txt
pdfplumber==0.10.0
chromadb==0.4.15
openai>=1.0.0
fpdf2==2.7.7
python-dotenv>=0.21.0
tqdm>=4.65.0
langchain>=0.0.148
langchain-openai>=0.0.1
langchain-community>=0.1.0
chainlit>=1.0.0
streamlit>=1.32.0
qdrant-client>=0.3.0
llama-parse>=0.0.1
nest-asyncio>=1.5.6
torch>=2.0.0
sentence-transformers>=2.2.2
langgraph>=0.1.0